Authentication Resource - Improve security and fraud issues by using a JWT Token. See below:
Explanation:

After the Email/External Account Verification step succeeded, store all the User Unique data
such as Email, Vendor User Id, in a JWT - Verification Token, instead of regular fields
passed back and forth from the server to the client and back to ensure they don't get
modified.

For example a really easy way to hack this would be after verifying an email, to modify the email
in the actual registration submit, which then will create a user with the modified unverified email.
Same goes foe External Accounts and External user id.

With the JWT token, all the information is stored inside of the token thus not allowing any modification
after verification. By using the JWT token we also get the benefits of a stateless transaction,
thus avoiding any need to save transitory data in the DB.